Added list feature and fixed bugs
